Resolved the three both-sides-added files with the resolution from the
GPU-tested composite lab/flashnext-pool-native af41ce309; the resulting
tree is identical to that composite (B 48.92 tok/s, 48 GiB audit and
48-layer init verification PASS, results/expert-tier-global-native-copy64-20260907).

+++ vllm/_lab_expert_tier/staging.py
@@ -300,7 +300,8 @@ def _copy_kernel():
         count = tl.load(count_ptr)
         if slot >= count:
             return
-        src_row = tl.load(index_ptr + slot)
+        src_row = tl.load(index_ptr + slot).to(tl.int64)
+        slot = slot.to(tl.int64)
         offsets = chunk * BLOCK + tl.arange(0, BLOCK)
         if which == 0:
             mask = offsets < bytes0
